fixed memory leak when returning nan. changed input from string to equation. fixed issue when computing result for pow and mult
i)as tmp used malloc there was a problem as it was not freed when returning nan. now tmp is a fixed size array
ii)now using the object equation as input for use with arduino calc keyboard. changed brackets algorithm to include operators like sin,cos,... as opening bracket
iii) fixed an issue when computing result for *,/,^ where  previously, if multindex==0, numbers.at(-1) would have been accessed this is no longer allowed